added mergeConfigs.go
